+++ cratedb_toolkit/__init__.py
@@ -9,3 +9,11 @@ try:
     __version__ = version(__appname__)
 except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
+
+from .config import preconfigure
+
+preconfigure()
+
+from .api import ManagedCluster  # noqa: E402, F401
+from .config import configure  # noqa: E402, F401
+from .model import InputOutputResource, TableAddress  # noqa: E402, F401