+++ cratedb_toolkit/cli.py
@@ -9,6 +9,7 @@ from .cluster.cli import cli as cloud_cli
 from .io.cli import cli as io_cli
 from .job.cli import cli_list_jobs
 from .shell.cli import cli as shell_cli
+from .util.setting import init_dotenv
 from .wtf.cli import cli as wtf_cli
 
 
@@ -18,6 +19,7 @@ from .wtf.cli import cli as wtf_cli
 @click.version_option()
 @click.pass_context
 def cli(ctx: click.Context, verbose: bool, debug: bool):
+    init_dotenv()
     return boot_click(ctx, verbose, debug)