+++ tests/io/test_import.py
@@ -28,6 +28,7 @@ def test_import_csv_dask(cratedb, dummy_csv):
     """
     Invoke convenience function `import_csv_dask`, and verify database content.
     """
+    pytest.importorskip("dask")
     result = cratedb.database.import_csv_dask(filepath=dummy_csv, tablename="foobar")
     assert result is None
 
@@ -42,6 +43,7 @@ def test_import_csv_dask_with_progressbar(cratedb, dummy_csv):
     This time, use `progress=True` to make Dask display a progress bar.
     However, the code does not verify it.
     """
+    pytest.importorskip("dask")
     result = cratedb.database.import_csv_dask(filepath=dummy_csv, tablename="foobar", progress=True)
     assert result is None
 
@@ -50,9 +52,9 @@ def test_import_csv_dask_with_progressbar(cratedb, dummy_csv):
     assert result == [(2,)]
 
 
-@pytest.mark.skip("Does not work. Why?")
+@pytest.mark.skip("Does not work. Q: Why? A: Response mocking? Q: And now? A: Just patch the low-level functions!")
 @responses.activate
-def test_import_cloud_file(tmp_path, caplog, cloud_cluster_mock):
+def test_import_cloud_file(tmp_path, caplog, mock_cloud_import):
     """
     CLI test: Invoke `ctk load table ...` for a CrateDB Cloud Import, from a local file.
     """
@@ -86,7 +88,7 @@ def test_import_cloud_file(tmp_path, caplog, cloud_cluster_mock):
 
 
 @responses.activate
-def test_import_cloud_url(caplog, cloud_cluster_mock):
+def test_import_cloud_url(caplog, mock_cloud_import):
     """
     CLI test: Invoke `ctk load table ...` for a CrateDB Cloud Import, from a URL.
     """