+++ CHANGES.md
@@ -4,6 +4,7 @@
 ## Unreleased
 - IO: Added the `if-exists` query parameter by updating to influxio 0.4.0.
 - Rockset: Added CrateDB Rockset Adapter, a HTTP API emulation layer
+- Cloud API: SDK and CLI for CrateDB Cloud Cluster and Import APIs
 
 ## 2024/06/18 v0.0.14
 - Add `ctk cfr` and `ctk wtf` diagnostics programs